Stop following redirects in Pegass session check
Context
-------

When performing a request to Pegass, the application checks whether the session is still active.

This check consists in calling a simple Pegass API and checking the http response.

Problem
-------

Even when the session is expired, the API check returns an HTTP 200 status code.

This is because Pegass automatically redirects the caller to the authentication provider.

Fix
---

Disable following redirects in HTTP calls when checking session status.